langchain-agents/requirements.txt
langchain
openai
replicate
python-dotenv
